sale-question-text-to-sql/BE/requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
openpyxl
langchain
langchain-openai
langchain-community
openai
python-dotenv
pydantic